changes in unit tests
